+++ packages/assets-controllers/src/index.ts
@@ -142,7 +142,7 @@ export {
   SUPPORTED_CHAIN_IDS,
   getNativeTokenAddress,
 } from './token-prices-service';
-export { searchTokens } from './token-service';
+export { searchTokens, getTrendingTokens } from './token-service';
 export { RatesController, Cryptocurrency } from './RatesController';
 export type {
   RatesControllerState,
@@ -224,9 +224,11 @@ export {
   calculateBalanceChangeForAccountGroup,
 } from './balances';
 export type {
+  AssetsByAccountGroup,
   AccountGroupAssets,
   Asset,
   AssetListState,
 } from './selectors/token-selectors';
 export { selectAssetsBySelectedAccountGroup } from './selectors/token-selectors';
 export { createFormatters } from './utils/formatters';
+export type { SortTrendingBy, TrendingAsset } from './token-service';