+++ packages/eth-json-rpc-middleware/test/util/helpers.ts
@@ -1,7 +1,7 @@
 import { PollingBlockTracker } from '@metamask/eth-block-tracker';
 import {
   providerFromEngine,
-  type SafeEventEmitterProvider,
+  type InternalProvider,
 } from '@metamask/eth-json-rpc-provider';
 import {
   JsonRpcEngine,
@@ -264,7 +264,7 @@ export function expectProviderRequestNotToHaveBeenMade(
  * you can make assertions on the method later, if you like).
  */
 export function stubProviderRequests(
-  provider: SafeEventEmitterProvider,
+  provider: InternalProvider,
   stubs: ProviderRequestStub<any, Json>[],
 ) {
   const remainingStubs = klona(stubs);